Remove all unused code
