--- disasm_arm.go
@@ -1,14 +1,0 @@
-package hookingo
-
-import (
-	"golang.org/x/arch/arm/armasm"
-)
-
-func analysis(src []byte) (inf info, err error) {
-	inst, err := armasm.Decode(src, armasm.ModeARM)
-	if err != nil {
-		return
-	}
-	inf.length = inst.Len
-	return
-}
