--- hook_ppc64le.go
@@ -1,15 +1,0 @@
-package hookingo
-
-import (
-	"errors"
-	"runtime"
-)
-
-const (
-	jumperSize = 0
-)
-
-func applyHook(from, to uintptr) (*hook, error) {
-
-	return nil, errors.New("TODO " + runtime.GOARCH + "Not supported")
-}
